Fixed several issues: 1) An error was being thrown when changing the lisp pipe type without having a path set. 2) Updated the SciLexer dlls Lisp Lexer to include "folding regions" capability. Now ;{ and ;} will act as "#region" and "#endregion" do in C#. 3) Updated the SciLexer dlls Lisp Lexer to allow disabling of compact folding (I don't like compact folding). 4) Updated the IDE to use the new features of the update Lisp Lexer. 5) The Lisp pipe now keeps the caret in view when resizing. 6) Disabling/Enabling folding should now work properly. 7) Updated installer versions to 1.1.0. 8) Changed the assembly major and minor version numbers to match that of the installers.
git-svn-id: https://lispidedotnet.svn.codeplex.com/svn@13777 a83551a4-30f6-4d81-a974-c6ced450ddbf

+++ LispIDEdotNet/Forms/LispPipe.cs
@@ -77,6 +77,8 @@ namespace LispIDEdotNet.Forms
             InitializeComponent();
 
             this.lispProcess = new Process();
+            this.Disposed += (sender, e) => CloseLisp();
+            this.Closing += (sender, e) => CloseLisp();
         }
 
         #region Methods
@@ -98,10 +100,9 @@ namespace LispIDEdotNet.Forms
 
         protected void OpenLisp()
         {
-            if(started)
-            {
-                CloseLisp();
-            }
+            // Close the lisp process if one has already
+            // been started.
+            CloseLisp();
 
             if(String.IsNullOrEmpty(this.LispPath))
             {
@@ -149,34 +150,38 @@ namespace LispIDEdotNet.Forms
 
         protected void CloseLisp()
         {
-            if (this.output != null)
+            if (started)
             {
-                this.output.CancelOperation();
-                this.output.Close();
+                if (this.output != null)
+                {
+                    this.output.CancelOperation();
+                    this.output.Close();
+                }
+
+                this.lispProcess.CancelErrorRead();
+                this.lispProcess.Exited -= lispProcess_Exited;
+                this.lispProcess.ErrorDataReceived -= lispProcess_ErrorDataReceived;
+
+                // We need to close the process if it hasn't exited
+                if (!this.lispProcess.HasExited)
+                {
+                    // Try and close the process cleanly
+                    this.SendCommand("(quit)" + Environment.NewLine + "(bye)" + Environment.NewLine + "(exit)" +
+                                     Environment.NewLine);
+
+                    // If that didnt work, force kill the process
+                    // Note: GCL (and possibly other REPLs) run off a .bat file. Killing the process
+                    // only kills the cmd.exe process and not the actual REPL. However, exiting
+                    // LispIDE.Net does kill the actual REPL process. I'm not sure how to force
+                    // kill the child process.
+                    if (!this.lispProcess.WaitForExit(100))
+                        this.lispProcess.Kill();
+                }
+
+                this.lispProcess.Close();
+
+                started = false;
             }
-
-            this.lispProcess.CancelErrorRead();
-            this.lispProcess.Exited -= lispProcess_Exited;
-            this.lispProcess.ErrorDataReceived -= lispProcess_ErrorDataReceived;
-            
-            // We need to close the process if it hasn't exited
-            if (!this.lispProcess.HasExited)
-            {
-                // Try and close the process cleanly
-                this.SendCommand("(quit)" + Environment.NewLine + "(bye)" + Environment.NewLine + "(exit)" + Environment.NewLine);
-
-                // If that didnt work, force kill the process
-                // Note: GCL (and possibly other REPLs) run off a .bat file. Killing the process
-                // only kills the cmd.exe process and not the actual REPL. However, exiting
-                // LispIDE.Net does kill the actual REPL process. I'm not sure how to force
-                // kill the child process.
-                if(!this.lispProcess.WaitForExit(100))
-                    this.lispProcess.Kill();
-            }
-
-            this.lispProcess.Close();
-           
-            started = false;
         }
 
         public virtual void Configure(ScintillaNet.Configuration.Configuration config)
@@ -229,7 +234,17 @@ namespace LispIDEdotNet.Forms
         private void LispPipe_Load(object sender, EventArgs e)
         {
             this.Scintilla.IsReadOnly = true;
+            this.Scintilla.SizeChanged += new EventHandler(Scintilla_SizeChanged);
             this.ScintillaBuffer.BufferReady += new EventHandler(ScintillaBuffer_BufferReady);
+        }
+
+        private void Scintilla_SizeChanged(object sender, EventArgs e)
+        {
+            // We want the caret to be visibale after a size change.
+            // This solution simply scrolls to the caret position so that it's the last
+            // line in the view (if it wasn't in view before). This may not be the 
+            // desired behavior, so I'll have to test it out and see.
+            this.Scintilla.Scrolling.ScrollToCaret();
         }
 
         private void LispPipe_FormClosed(object sender, FormClosedEventArgs e)
